Fix memory accounting if replacing internal node w/ larger one goes over limit, part 1
During insertion, if we do not find key value in the tree, create a
leaf node without going over the memory limit, but then we need to
expand an internal node to the next node type, and if that allocation
goes over limit, then the leaf node is deallocated without reducing
the memory use counter. Fix accordingly, add a test.