Fix memory leaks on replacing smaller nodes with larger ones
Fix by making larger node factory methods and constructor take rvalue references
to unique pointers to smaller nodes instead of const references (and relying on
the caller to free the smaller nodes, which was node done). This matches what
the opposite case of creating a smaller node out of a larger one does.

This is a regression from the node_ptr conversion from unique_ptr to raw
pointers and fixes the newly-uncovered Valgrind errors.