Introduce unodb::detail::tree_depth value type
Previously tree_depth was an alias of unsigned. We want to introduce an
invariant for these values-bound the range-but adding asserts at each call site
results in duplicated code and it's too easy to miss relevant call sites.

Thus wrap the number in a small helper class where the invariant can be checked
locally.

Add a couple of other semi-related asserts too.