Trim the heap at the end of each micro benchmark
If a micro benchmark completes that used a lot of heap memory, then the next
test start will go through long reclamation on first allocation or deallocation,
throwing the benchmark numbers off, with glibc allocator. Ask for this work
explicitly at the end of each benchmark, in a non-timed section.

TODO: numbers if this ok on master not only on direct-heap.