Add sentinel children pointers to internal nodes
This allows removing one more branch from inode_4::find_child and prepares the
rest of the internal nodes for the same.

In order to avoid mixing zero- and one- based indexing in children arrays
depending on whether the sentinel is counted or not, introduce union
children_with_sentinel, which has overlapping children arrays both including and
not including the sentinel.

Replace pointer with reference in inode::find_result_type now that the result of
find_child can never be a nullptr itself.

With this change, sizeof(inode_4) goes from 48 bytes to 56 bytes.