Fix the lifetime of olc_db::get result
Previously, the olc_db::get went through QSBR quiescent state itself, while
returning a value_view, that is, a pointer to a value in the tree. Which means
that passing through a quiescent state is incorrect.

Fix by removing all the quiescent state calls in olc_db and adding them in
callers as required. To try to exclude this bug class, introduce new helper
types: qsbr_ptr - which is an active pointer to QSBR-managed shared data,
asserting in the debug build that quiescent state can only be passed with zero
active pointers, and qsbr_value_view - which is a gsl::span/std::span but build
on qsbr_ptr, and return that from olc_db::get.

Add -Wno-unused-macros to .clangd to further workaround
https://github.com/clangd/clangd/issues/363

Do assorted suggested minor cland-tidy and IWYU cleanups.