Add qsbr_ptr support to the QSBR fuzzer
Add taking and releasing active pointers to the QSBR fuzzer actions.

Try to increase fuzzing depth by picking another action to do when the state
does not allow performing already-selected one instead of dropping it and
proceeding with the next one.

Extract DeepSTate_ContainerIndex helper, refactor the ART fuzzer to use it too.