Fix locking and lifetime for unodb::mutex_db query results
The problem was the mutex_db::get returned views into values with lock released.
Fix by returning a pair of search result and locked mutex instead, which then is
released after inspecting the search result as required. This might be not the
best design but mutex_db exists only as a concurrency baseline anyway.

Introduce a helper key_found in all the ART classes that checks whether a key
was found in a given search result, and, in the case of mutex_db, releases the
mutex.

At the same time move all the get_result types into their corresponding ART
classes.