Require commit call in every success path of unique_write_lock_obseleting_guard
Previously, destructor checked std::uncaught_exceptions and assumed that
destruction while an exception is being thrown is abort, otherwise commit. This
did not work out due to guard lifetime extending past the end of its variable
scope due to copy elision, and explicit commit (or abort) is required on every
path. Thus destructor is now free to assume that if the lock is still locked,
it's an abort and remove all the std::uncaught_exceptions business.

Ad the same time add a couple of asserts to the move constructor.