Fix parallel_get not searching the last range and searching the first range twice
Due to off-by-one error, the last key range to search was not assigned to any
thread, and the first range was searched by both the first spawned thread and
the main thread.

Fix accordingly, at the same time refactor all parallel benchmarks to use a
do_parallel_test helper, removing the duplicated code and fixing minor
differences on the way.