Make qsbr::assert_idle locking
As pointed out by Sonatype Lift Infer checker, qsbr::assert_idle accesses fields
without locking. While it is always meant to be invoked in a single-threaded
context, the purpose of the asserts is to catch bugs, including when they are
called from multi-threaded context. Thus make it locking, and introduce a
private member function assert_idle_locked to be used when the lock is already
acquired.