Implement CMake option FATAL_WARNINGS
By default, compiler warnings used to be fatal. While this is good for me as a
maintainer, it is not good for users. Thus make it optionally enabled by a new
CMake option FATAL_WARNINGS. Add this option to CI/CD jobs, CMake presets. Also
add a README.md note.