Lock-free QSBR step 1: move q-states-seen from global to per-thread data structure
In the old design, the per-thread quiescent states since the last epoch change
were in the shared QSBR data structure. Make that per-thread to reduce locking.
To make it work, each thread now tracks the last global epoch it has seen to
determine whether the epoch advanced since the last quiescent state for this
thread.

This allows reducing the shared data structure from a map from thread ids to q
states seen to a set of threads ids, which will be removed next.

Other changes:
- Introduce UNODB_DETAIL_RELEASE_STATIC macro for methods that are static in
  release build only
- Likewise UNODB_DETAIL_DEBUG_CONST for debug build-only const
- Move the epoch_type type alias out of qsbr class to unodb namespace and rename
  it to qsbr_epoch