Fix QSBR epoch change counter race
Previously, if thread unregistration would try to advance the epoch, the change
counter would be bumped before the attempt – even if it failed. To fix, split
qsbr::bump_epoch_change_count method out of epoch_change_update_requests. Rename
the latter to epoch_change_barrier_and_handle_orphans. Remove incorrect assert
in it which tried to assert orphaned request deallocation epoch being in the
past – it is impossible to have it correct with a two-bit epoch counter because
the orphaned request epoch lag is unbounded.