WIP Implement lock-free statistics and use it for QSBR
Implement lock-free statistics that use DWCAS to atomically update count and
mean. Additional optional stats of maximum and variance are implemented on the
top of that. Replace QSBR stats from Boost.Accumulator to this. The
implementation requires for x86_64, thus add -mcx16 compilation flag in the case
of SSE4.2/AVX implementation.