Fix including UnoDB into other projects
Do not define global debugging checks _GLIBCXX_DEBUG and _GLIBCXX_DEBUG_PEDANTIC
unconditionally, because they require the whole program to be built either with,
either without the define.

Add a new CMake option STANDALONE which indicates that UnoDB is being built
separately, and it is OK to enable such checks.