Force inline optimistic_lock::version_type::is_obsolete.
LLVM 14-17 and possibly later versions generate a call  to outline version from
optimistic_lock::try_lock in release build with UBSan. That same method is
apparently miscompiled in that its loop only checks whether the lock is free but
never if it's obsolete, resulting in hangs. Forcing to inline seems to make that
issue to go away too.