WIP - test case with 2 leaves under an N4 fails validation.
It seems likely that the tree is built incorrectly.  The N4 has a
key_prefix equal to the key on one of the leaves. That means that
there is no key_byte along which we can descend to that leaf.  This
shows up as an runtime error when we attempt to validate the entries
in the index using get().